reorganized meepmeep1x files
