+++ MeepMeep1x/src/main/java/com/example/meepmeep1x/decembercomp/MeepMeepBlueSpecimenAuto.java
@@ -1,4 +1,4 @@
-package com.example.meepmeep1x;
+package com.example.meepmeep1x.decembercomp;
 
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.Vector2d;
